AdaptiveHierarchicalRegularBinning

Documentation for AdaptiveHierarchicalRegularBinning.

AdaptiveHierarchicalRegularBinning.bit_interleave

AdaptiveHierarchicalRegularBinning.bit_interleave

AdaptiveHierarchicalRegularBinning.bit_interleave

AdaptiveHierarchicalRegularBinning.bit_space

AdaptiveHierarchicalRegularBinning.bit_space_mask

AdaptiveHierarchicalRegularBinning.countsortperm!

AdaptiveHierarchicalRegularBinning.morton

AdaptiveHierarchicalRegularBinning.pcountsortperm!

AdaptiveHierarchicalRegularBinning.translate_scale_vals

AdaptiveHierarchicalRegularBinning.bit_interleave — Method

bit_interleave(W[; dims])

Interleaves all unsigned integers along a dimention.

Arguments

W: The matrix

dims: The dimention. Defaults to 1

AdaptiveHierarchicalRegularBinning.bit_interleave — Method

bit_interleave([M, ]W)

Interleaves an array of unsigned integers.

Arguments

M: The method with which to interleave

W: The array

Examples

julia> bit_interleave([0x00FF, 0x000F]) |> bitstring
"0101010111111111"

julia> bit_interleave([0x000F, 0x00FF]) |> bitstring
"1010101011111111"

julia> bit_interleave([0x0080, 0x0001]) |> bitstring
"0100000000000010"

julia> bit_interleave([0x0001, 0x0080]) |> bitstring
"1000000000000001"

AdaptiveHierarchicalRegularBinning.bit_interleave — Method

bit_interleave(Ws)

Interleaves all unsigned integers.

AdaptiveHierarchicalRegularBinning.bit_space — Method

bit_space(M, w, n)

Evenly space out bits.

Arguments

M: The method to space out the bits

w: The word to space out

n: The amount of padding

Examples

julia> bit_space(Brute, 0x00ABCDEF, 0) |> bitstring
"00000000101010111100110111101111"

julia> bit_space(Brute, 0x00ABCDEF, 1) |> bitstring
"10100000101000101010100010101010"

julia> bit_space(Brute, 0x00ABCDEF, 2) |> bitstring
"00000100100100100000100100100100"

julia> bit_space(Brute, 0x00ABCDEF, 3) |> bitstring
"10001000100000001000100010001000"

julia> bit_space(Pdep, 0x00ABCDEF, 2) == bit_space(Brute, 0x00ABCDEF, 2)
true

AdaptiveHierarchicalRegularBinning.bit_space_mask — Method

bit_space_mask(_, _)

Generates a spaced mask.

Arguments

T: The resulting type

n: The amount of padding

Examples

julia> bit_space_mask(UInt8, 0) |> bitstring
"11111111"

julia> bit_space_mask(UInt8, 1) |> bitstring
"01010101"

julia> bit_space_mask(UInt8, 2) |> bitstring
"01001001"

julia> bit_space_mask(UInt8, 3) |> bitstring
"00010001"

AdaptiveHierarchicalRegularBinning.countsortperm! — Method

countsortperm!(p, v, C)

Sequential count sort permutation. Overwrite vector p with the permutation that puts vector of values v[p] in non decreasing order.

p: preallocated output vector of integers to receive the permutation

v: input vector of integer values that determine order. The unique values of v must be in the range 0:m where m << length(v)

C: preallocated workspace of size at least m+1

AdaptiveHierarchicalRegularBinning.morton — Method

morton(p, L) -> Any

Spatially encode a single point 0 ≤ p < 1 for L levels

AdaptiveHierarchicalRegularBinning.pcountsortperm! — Method

pcountsortperm!(p::Vector{Int64}, v::Vector{Int64}, C::Matrix{Int64})

Multithreaded count sort permutation. Overwrite vector p with the permutation that puts vector of values v[p] in non decreasing order.

p: preallocated output vector of integers to receive the permutation

v: input vector of integer values that determine order. The unique values of v must be in the range 0:m where m << length(v)

C: preallocated workspace of size at least (m+1+8:np+1), where np is the number of threads.

AdaptiveHierarchicalRegularBinning.translate_scale_vals — Method

translate_scale_vals(p; center, dims) -> Tuple{Any, Any}

Find the displacement vector δ and scale scalar σ parameters that will transform the cloud points p to be within a unit box [0,1),

0 ≤ σ (p - δ) < 1

p: cloud point coordinates

center: flag to center the cloud by dividing the slack evenly, default false

dims : dimension of p expressing single point coordinates,

i.e. if the i^th point is p[i,:], then set dims=2.