AdaptiveHierarchicalRegularBinning

Documentation for AdaptiveHierarchicalRegularBinning.

AdaptiveHierarchicalRegularBinning.NodeInfo

AdaptiveHierarchicalRegularBinning.SpatialTree

AdaptiveHierarchicalRegularBinning.TreeInfo

AdaptiveHierarchicalRegularBinning.ahrb

AdaptiveHierarchicalRegularBinning.box2boxdist

AdaptiveHierarchicalRegularBinning.compute_centers_sidelengths!

AdaptiveHierarchicalRegularBinning.copy_final!

AdaptiveHierarchicalRegularBinning.find_rec_leaves

AdaptiveHierarchicalRegularBinning.getcontext

AdaptiveHierarchicalRegularBinning.point2boxdist

AdaptiveHierarchicalRegularBinning.select_pingpong_buffers_subdivision

AdaptiveHierarchicalRegularBinning.setcontext!

AdaptiveHierarchicalRegularBinning.specialprioritymultilevelinteractions

AdaptiveHierarchicalRegularBinning.NodeInfo — Type

struct NodeInfo

Represents a node of the tree.

Fields

lo: The start index of the represented group.

hi: The stop index of the represented group.

dpt: The depth of the node

nidx: The index of the node.

pidx: The index of the parent node.

AdaptiveHierarchicalRegularBinning.SpatialTree — Type

struct SpatialTree{T, E, C, GTC}

Represents the tree.

Fields

T: The element type of the points.

E: The element type of the encoded points.

B: The bit length of each bit group in the encoded space.

D: The leading dimension.

info: The tree information.

nidx: Index to the current root of the tree.

AdaptiveHierarchicalRegularBinning.TreeInfo — Type

struct TreeInfo{T, E, C, GTC}

Represents the tree information.

Fields

T: The element type of the points.

E: The element type of the encoded points.

B: The bit length of each bit group in the encoded space.

D: The leading dimension.

points: The cloud of points that the tree spans.

encoded: The encoded cloud of points.

perm: The permutation of the points.

scale: The original scale of the dimensions of the points.

offset: The per dimension offset of the points.

nodes: Per node information.

children: Per node children.

maxdepth: The maximum depth of the tree.

maxpoints: Small threshold.

AdaptiveHierarchicalRegularBinning.ahrb — Function

ahrb(V) -> SpatialTree
ahrb(V, maxdepth) -> SpatialTree
ahrb(
    V,
    maxdepth,
    maxpoints;
    QT,
    ctxtype,
    gtctype,
    method,
    lstep
) -> SpatialTree

Generate an adaptive hierarchical regular binning tree structure.

Arguments

V: A matrix of data points, where each column represents a data point and each row represents a feature.

maxdepth: An integer that specifies the maximum depth of the tree. Default is ceil(Int, log2(size(V,2))).

maxpoints: An integer that specifies the maximum number of points in a leaf node. Default is ceil(Int, (size(V,2))^(1/4)).

QT: (Optional) The type of the vector R used to store the spatial encoding of the data points. Defaults to UInt.

ctxtype: (Optional) The type of the context information stored at each node of the tree. Defaults to Nothing.

gtctype: (Optional) The type of the global context information stored alongside the tree. Defaults to Nothing.

lstep: (Optional) The number of levels to process in each block iteration. Defaults to 2.

method: (Optional) The method to use for the spatial encoding. Defaults to "block-ecp". Possible values are:

"block-ecp": Use the block ECP version.

"fixed-length": Use the fixed length encoding.

Returns

The generated tree structure.

Example

julia> V = rand(6, 10_000);

julia> tree = ahrb(V, 5, 10)
SpatialTree: 
Matrix{Float64}(6,10000) points
3805 nodes, 3740 leaves and max depth 2

AdaptiveHierarchicalRegularBinning.box2boxdist — Method

box2boxdist(c1, h1, c2, h2)

box2boxdist(node1, node2)

Compute the distance between two boxes. c denotes the center of the box, h denotes the half-width of the box.

AdaptiveHierarchicalRegularBinning.compute_centers_sidelengths! — Method

compute_centers_sidelengths!(
    centers::Array{Array{V<:Real, 1}, 1},
    sidelengths::Array{V<:Real, 1},
    nodes::Vector{NodeInfo},
    R0::Array{Tu<:Unsigned, 1},
    currentdepth::Int64,
    d::Int64
)

Compute the center and sidelength of each node in the tree.

AdaptiveHierarchicalRegularBinning.copy_final! — Method

copy_final!(
    R0::AbstractVector,
    R1::AbstractVector,
    data_0or1::Vector{Int8}
)

Copy the codes and the permutation vector between the ping-pong buffers for the final result

AdaptiveHierarchicalRegularBinning.find_rec_leaves — Method

find_rec_leaves(
    nodes::Vector{NodeInfo},
    children::Vector{Vector{Int64}},
    maxpts::Int64
) -> Vector{Int64}

Find leaves that are above the maximum number of points per leaf and we should recurse to subdivide them.

AdaptiveHierarchicalRegularBinning.getcontext — Method

Read context of a node

AdaptiveHierarchicalRegularBinning.point2boxdist — Method

point2boxdist(p, c, h)

point2boxdist(p, node)

Compute the distance between a point and a box. p denotes the point. c denotes the center of the box, h denotes the half-width of the box.

AdaptiveHierarchicalRegularBinning.select_pingpong_buffers_subdivision — Method

select_pingpong_buffers_subdivision(
    R0,
    V0,
    Pglb,
    R1,
    V1,
    Plcl,
    rng,
    data_0or1
) -> NTuple{6, Any}

Select which buffer to use for the ping-pong algorithm.

AdaptiveHierarchicalRegularBinning.setcontext! — Method

Set context of a node

AdaptiveHierarchicalRegularBinning.specialprioritymultilevelinteractions — Method

Specialized version of prioritymultilevelinteractions where t is constant leaf.