AdaptiveHierarchicalRegularBinning

Documentation for AdaptiveHierarchicalRegularBinning.

AdaptiveHierarchicalRegularBinning.CountSortDetails

AdaptiveHierarchicalRegularBinning.CountSortDetails

AdaptiveHierarchicalRegularBinning.CountSortDetails

AdaptiveHierarchicalRegularBinning.RadixSortDetails

AdaptiveHierarchicalRegularBinning.RadixSortDetails

AdaptiveHierarchicalRegularBinning.bit_interleave

AdaptiveHierarchicalRegularBinning.bit_interleave

AdaptiveHierarchicalRegularBinning.bit_interleave

AdaptiveHierarchicalRegularBinning.bit_space

AdaptiveHierarchicalRegularBinning.bit_space_mask

AdaptiveHierarchicalRegularBinning.bitlen

AdaptiveHierarchicalRegularBinning.bitmask

AdaptiveHierarchicalRegularBinning.countsort_par_impl!

AdaptiveHierarchicalRegularBinning.countsort_seq_impl!

AdaptiveHierarchicalRegularBinning.depth

AdaptiveHierarchicalRegularBinning.high

AdaptiveHierarchicalRegularBinning.leaddim

AdaptiveHierarchicalRegularBinning.low

AdaptiveHierarchicalRegularBinning.next

AdaptiveHierarchicalRegularBinning.next

AdaptiveHierarchicalRegularBinning.quantize

AdaptiveHierarchicalRegularBinning.radixsel

AdaptiveHierarchicalRegularBinning.radixsort_par_par_impl!

AdaptiveHierarchicalRegularBinning.radixsort_par_seq_impl!

AdaptiveHierarchicalRegularBinning.radixsort_seq_seq_impl!

AdaptiveHierarchicalRegularBinning.spatial_encode!

AdaptiveHierarchicalRegularBinning.translate_scale_vals

Base.eltype

Base.length

AdaptiveHierarchicalRegularBinning.CountSortDetails — Type

struct CountSortDetails{B, D}

Contains all of the details needed for the countsort_xxx_impl! algorithms.

Fields

B: The length of the bit mask to use.

D: The leading dimension.

lo: Lower array bound.

hi: Higher array bound.

l: The level of recursion.

AdaptiveHierarchicalRegularBinning.CountSortDetails — Method

CountSortDetails(rsd)

Gets a CountSortDetails object from a RadixSortDetails object.

AdaptiveHierarchicalRegularBinning.CountSortDetails — Method

CountSortDetails(bitlen, lo, hi; dims)

Constructs a new CountSortDetails object.

Arguments

bitlen: The bit length of the bit mask used by the algorithm.

lo: Lower array bound.

hi: Higher array bound.

Keyword Arguments

dims: The leading dimension.

AdaptiveHierarchicalRegularBinning.RadixSortDetails — Type

struct RadixSortDetails{B, D}

Contains all of the details needed for the radixsort_xxx_xxx_impl! algorithms.

Fields

B: The length of the bit mask to use.

D: The leading dimension.

csd: The underlying CountSortDetails object.

par_th: Threshold for parallel-to-parallel execution.

seq_th: Threshold for parallel-to-sequential execution.

sml_th: Threshold for skiping small arrays.

dpt_th: recursion threshold.

AdaptiveHierarchicalRegularBinning.RadixSortDetails — Method

RadixSortDetails(bitlen, lo, hi; dims, par_th, seq_th, sml_th, dpt_th)

Constructs a new RadixSortDetails object.

Arguments

bitlen: The bit length of the bit mask used by the algorithm.

lo: Lower array bound.

hi: Higher array bound.

Keyword Arguments

par_th: Threshold for parallel-to-parallel execution.

seq_th: Threshold for parallel-to-sequential execution.

sml_th: Threshold for skiping small arrays.

dpt_th: recursion threshold.

AdaptiveHierarchicalRegularBinning.bit_interleave — Method

bit_interleave(W[; dims])

Interleaves all unsigned integers along a dimention.

Arguments

W: The matrix

dims: The dimention. Defaults to 1

AdaptiveHierarchicalRegularBinning.bit_interleave — Method

bit_interleave([M, ]W)

Interleaves an array of unsigned integers.

Arguments

M: The method with which to interleave

W: The array

Examples

julia> bit_interleave([0x00FF, 0x000F]) |> bitstring
"0101010111111111"

julia> bit_interleave([0x000F, 0x00FF]) |> bitstring
"1010101011111111"

julia> bit_interleave([0x0080, 0x0001]) |> bitstring
"0100000000000010"

julia> bit_interleave([0x0001, 0x0080]) |> bitstring
"1000000000000001"

AdaptiveHierarchicalRegularBinning.bit_interleave — Method

bit_interleave(Ws)

Interleaves all unsigned integers.

AdaptiveHierarchicalRegularBinning.bit_space — Method

bit_space(M, w, n)

Evenly space out bits.

Arguments

M: The method to space out the bits

w: The word to space out

n: The amount of padding

Examples

julia> bit_space(Brute, 0x00ABCDEF, 0) |> bitstring
"00000000101010111100110111101111"

julia> bit_space(Brute, 0x00ABCDEF, 1) |> bitstring
"10100000101000101010100010101010"

julia> bit_space(Brute, 0x00ABCDEF, 2) |> bitstring
"00000100100100100000100100100100"

julia> bit_space(Brute, 0x00ABCDEF, 3) |> bitstring
"10001000100000001000100010001000"

julia> bit_space(Pdep, 0x00ABCDEF, 2) == bit_space(Brute, 0x00ABCDEF, 2)
true

AdaptiveHierarchicalRegularBinning.bit_space_mask — Method

bit_space_mask(_, _)

Generates a spaced mask.

Arguments

T: The resulting type

n: The amount of padding

Examples

julia> bit_space_mask(UInt8, 0) |> bitstring
"11111111"

julia> bit_space_mask(UInt8, 1) |> bitstring
"01010101"

julia> bit_space_mask(UInt8, 2) |> bitstring
"01001001"

julia> bit_space_mask(UInt8, 3) |> bitstring
"00010001"

AdaptiveHierarchicalRegularBinning.bitlen — Method

Gets the bit length the object.

AdaptiveHierarchicalRegularBinning.bitmask — Method

bitmask(csd)

Gets the bit mask the object.

AdaptiveHierarchicalRegularBinning.countsort_par_impl! — Method

countsort_par_impl!(Va, Ra, Ia, Vb, Rb, Ib, C, csd)

Parallel countsort.

Remarks

Inputs Va, Ra, Ia are not altered.

Arguments

Va: The cloud of points.

Ra: The morton transformation for each point of Va

Ia: The current permutation of Va.

Vb: A matrix to store the sorted version of Va.

Rb: A vector to store the sorted version of Ra.

Ib: A vector to store the resulting permutation.

C: A matrix acting as a working space for the countsort algorithm.

csd: The CountSortDetails object used to configure the algorithm.

AdaptiveHierarchicalRegularBinning.countsort_seq_impl! — Method

countsort_seq_impl!(Va, Ra, Ia, Vb, Rb, Ib, C, csd)

Sequential countsort.

Remarks

Inputs Va, Ra, Ia are not altered.

Arguments

Va: The cloud of points.

Ra: The morton transformation for each point of Va

Ia: The current permutation of Va.

Vb: A matrix to store the sorted version of Va.

Rb: A vector to store the sorted version of Ra.

Ib: A vector to store the resulting permutation.

C: A vector acting as a working space for the countsort algorithm.

csd: The CountSortDetails object used to configure the algorithm.

AdaptiveHierarchicalRegularBinning.depth — Method

depth(csd)

Gets the level of recursion of the object.

AdaptiveHierarchicalRegularBinning.high — Method

high(csd)

Gets the higher index of the object.

AdaptiveHierarchicalRegularBinning.leaddim — Method

leaddim(_)

Gets the leading dimension the object.

AdaptiveHierarchicalRegularBinning.low — Method

low(csd)

Gets the lower index of the object.

AdaptiveHierarchicalRegularBinning.next — Method

next(csd, lo, hi)

Constructs a new CountSortDetails object with a deeper level of recursion.

Arguments

lo: New lower array bound.

hi: New higher array bound.

AdaptiveHierarchicalRegularBinning.next — Method

next(rsd, lo, hi)

Constructs a new RadixSortDetails object with a deeper level of recursion.

Arguments

lo: New lower array bound.

hi: New higher array bound.

AdaptiveHierarchicalRegularBinning.quantize — Method

quantize(T, x, l)

Quantizes a single coordinate.

Arguments

T: The resulting type.

x: The input coordinate.

l: The levels of the quantizer.

AdaptiveHierarchicalRegularBinning.radixsel — Method

radixsel(csd, x)

Selects the corresponding radix.

Arguments

csd: The CountSortDetails object to base the selection on.

x: The value from which the radix is computed.

AdaptiveHierarchicalRegularBinning.radixsort_par_par_impl! — Method

radixsort_par_par_impl!(Va, Ra, Ia, Vb, Rb, Ib, P, rsd, allocator)

Parallel-to-Parallel radixsort.

Arguments

Va: The cloud of points.

Ra: The morton transformation for each point of Va

Ia: The current permutation of Va.

Vb: An auxiliary matrix to store the partially sorted version of Va.

Rb: An auxiliary vector to store the partially sorted version of Ra.

Ib: An auxiliary vector to store the resulting permutation.

P: A ping-pong matrix that denotes where the latest results are.

rsd: The RadixSortDetails object used to configure the algorithm.

AdaptiveHierarchicalRegularBinning.radixsort_par_seq_impl! — Method

Parallel-to-Sequential radixsort.

Arguments

Va: The cloud of points.

Ra: The morton transformation for each point of Va

Ia: The current permutation of Va.

Vb: An auxiliary matrix to store the partially sorted version of Va.

Rb: An auxiliary vector to store the partially sorted version of Ra.

Ib: An auxiliary vector to store the resulting permutation.

P: A ping-pong matrix that denotes where the latest results are.

rsd: The RadixSortDetails object used to configure the algorithm.

AdaptiveHierarchicalRegularBinning.radixsort_seq_seq_impl! — Method

Sequential-to-Sequential radixsort.

Arguments

Va: The cloud of points.

Ra: The morton transformation for each point of Va

Ia: The current permutation of Va.

Vb: An auxiliary matrix to store the partially sorted version of Va.

Rb: An auxiliary vector to store the partially sorted version of Ra.

Ib: An auxiliary vector to store the resulting permutation.

P: A ping-pong matrix that denotes where the latest results are.

rsd: The RadixSortDetails object used to configure the algorithm.

AdaptiveHierarchicalRegularBinning.spatial_encode! — Method

spatial_encode!(R, V, l; dims, center)

Encodes a set of points using the morton encoding.

Arguments

R: The resulting vector.

V: The input matrix.

l: The levels of the encoding.

Keyword Arguments

dims: The leading dimension.

center: Whether to divide the slack evenly.

AdaptiveHierarchicalRegularBinning.translate_scale_vals — Method

translate_scale_vals(V; dims, center)

Compute displacement vector and the scale to transform the cloud to a unit hypercube.

Arguments

V: The input matrix.

Keyword Arguments

dims: The leading dimension.

center: Whether to divide the slack evenly.

Base.eltype — Method

eltype(csd)

The index type of the struct.

Base.length — Method

length(csd)

Gets the length of the object.